+++ android-movie/app/src/main/java/com/example/movie/MenuType.kt
@@ -1,0 +1,7 @@
+package com.example.movie
+
+enum class MenuType(val menu : Int) {
+    RESERVATION_DETAILS(0),
+    HOME(1),
+    SETTING(2)
+}
