+++ android-movie/app/src/main/res/layout/back_toolbar.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="56dp"
-    android:background="@color/forestgreen">
+    android:background="@color/forest_green">
 
     <ImageView
         android:id="@+id/btn_back"
