+++ android-movie/app/src/main/res/layout/fragment_bottom_sheet.xml
@@ -25,7 +25,7 @@
             android:layout_gravity="center_vertical"
             android:layout_marginStart="15dp"
             android:text="@string/theater_1"
-            android:textColor="@color/forestgreen"
+            android:textColor="@color/forest_green"
             android:textSize="18sp"
             android:textStyle="bold" />
 
@@ -53,7 +53,7 @@
             android:layout_gravity="center_vertical"
             android:layout_marginStart="15dp"
             android:text="@string/theater_2"
-            android:textColor="@color/forestgreen"
+            android:textColor="@color/forest_green"
             android:textSize="18sp"
             android:textStyle="bold" />
 
@@ -81,7 +81,7 @@
             android:layout_gravity="center_vertical"
             android:layout_marginStart="15dp"
             android:text="@string/theater_3"
-            android:textColor="@color/forestgreen"
+            android:textColor="@color/forest_green"
             android:textSize="18sp"
             android:textStyle="bold" />
