+++ android-movie/app/src/main/res/menu/menu_bottom_nav.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/reservation_details"
+        android:title="@string/menu_reservation_details"
+        android:icon="@drawable/list"/>
+    <item
+        android:id="@+id/home"
+        android:title="@string/menu_home"
+        android:icon="@drawable/home"/>
+    <item
+        android:id="@+id/setting"
+        android:title="@string/menu_setting"
+        android:icon="@drawable/setting"/>
+</menu>
